+++ src/BigRational.php
@@ -56,7 +56,7 @@ final readonly class BigRational extends BigNumber
     {
         if ($checkDenominator) {
             if ($denominator->isZero()) {
-                throw DivisionByZeroException::denominatorMustNotBeZero();
+                throw DivisionByZeroException::zeroDenominator();
             }
 
             if ($denominator->isNegative()) {